+++ bin/nanocld/changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Filter by kind, existing key and contains for data and metadata of a secret. [@anonkey](https://github.com/anonkey)
 - Filter by existing key and contains for data and metadata of a resource.
 - Before cargo, to run a command before creating the cargo.